requirements.txt
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
MouseInfo==0.1.3
pillow==10.4.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dotenv==1.0.1
python3-xlib==0.15
pytweening==1.2.0
pywhatkit==5.4
requests==2.32.3
soupsieve==2.6
urllib3==2.2.2
Werkzeug==3.0.4
wikipedia==1.4.0